Add files via upload (#17)
Removed String args[] from all the classes that werent main.

Experimental code for working with the search results
